+++ packages/datepicker/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.1.0](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-datepicker@7.0.20...@pluralsight/ps-design-system-datepicker@7.1.0) (2020-08-25)
+
+
+### Features
+
+* disabled cursor styles ([#1115](https://github.com/pluralsight/design-system/issues/1115)) ([94fd04c](https://github.com/pluralsight/design-system/commit/94fd04c1751231573c06ad590c037c2cf9397d02))
+
+
+
+
+
 ## [7.0.20](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-datepicker@7.0.19...@pluralsight/ps-design-system-datepicker@7.0.20) (2020-08-13)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-datepicker