+++ packages/note/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.0.6](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-note@7.0.5...@pluralsight/ps-design-system-note@7.0.6) (2020-08-25)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-note
+
+
+
+
+
+## [7.0.5](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-note@7.0.4...@pluralsight/ps-design-system-note@7.0.5) (2020-08-25)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-note
+
+
+
+
+
 ## [7.0.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-note@7.0.3...@pluralsight/ps-design-system-note@7.0.4) (2020-08-24)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-note